# Introduce SingleBytePrefixVersionWithStringBypass for faster string serialization