--- a/lib/paquito/single_byte_prefix_version.rb
+++ b/lib/paquito/single_byte_prefix_version.rb
@@ -3,8 +3,8 @@
 module Paquito
   class SingleBytePrefixVersion
     def initialize(current_version, coders)
-      @current_version = current_version
-      @coders = coders.transform_values { |c| Paquito.cast(c) }
+      @current_version = validate_version(current_version)
+      @coders = coders.transform_keys { |v| validate_version(v) }.transform_values { |c| Paquito.cast(c) }
       @current_coder = coders.fetch(current_version)
     end
 
@@ -22,6 +22,16 @@ module Paquito
         raise UnsupportedCodec, "Unsupported packer version #{payload_version}"
       end
       coder.load(payload.byteslice(1..-1))
+    end
+
+    private
+
+    def validate_version(version)
+      unless (0..255).cover?(version)
+        raise ArgumentError, "Invalid version #{version.inspect}, versions must be an integer between 0 and 255"
+      end
+
+      version
     end
   end
 end