… lib/paquito.rb