Fix: https://github.com/Shopify/paquito/issues/18

```
 === Read 100B ===
Warming up --------------------------------------
             marshal   230.578k i/100ms
             msgpack   242.521k i/100ms
              bypass   526.491k i/100ms
Calculating -------------------------------------
             marshal      2.210M (± 3.3%) i/s -     11.068M in   5.014646s
             msgpack      2.400M (± 2.6%) i/s -     12.126M in   5.057475s
              bypass      5.264M (± 0.8%) i/s -     26.325M in   5.001316s

Comparison:
             marshal:  2209529.3 i/s
              bypass:  5263899.3 i/s - 2.38x  (± 0.00) faster
             msgpack:  2399605.0 i/s - 1.09x  (± 0.00) faster

 === Write 100B ===
Warming up --------------------------------------
             marshal   236.093k i/100ms
             msgpack   244.387k i/100ms
              bypass   407.380k i/100ms
Calculating -------------------------------------
             marshal      2.287M (± 2.6%) i/s -     11.569M in   5.062488s
             msgpack      2.368M (± 2.1%) i/s -     11.975M in   5.060218s
              bypass      4.022M (± 1.9%) i/s -     20.369M in   5.065649s

Comparison:
             marshal:  2286724.8 i/s
              bypass:  4022499.9 i/s - 1.76x  (± 0.00) faster
             msgpack:  2367566.5 i/s - same-ish: difference falls within error


 === Read 10000B ===
Warming up --------------------------------------
             marshal    70.538k i/100ms
             msgpack   221.620k i/100ms
              bypass   526.303k i/100ms
Calculating -------------------------------------
             marshal    618.272k (± 8.4%) i/s -      3.104M in   5.055696s
             msgpack      2.219M (± 0.5%) i/s -     11.303M in   5.094451s
              bypass      5.253M (± 0.5%) i/s -     26.315M in   5.009325s

Comparison:
             marshal:   618272.2 i/s
              bypass:  5253358.9 i/s - 8.50x  (± 0.00) faster
             msgpack:  2218677.8 i/s - 3.59x  (± 0.00) faster

 === Write 10000B ===
Warming up --------------------------------------
             marshal    51.365k i/100ms
             msgpack    33.687k i/100ms
              bypass    65.035k i/100ms
Calculating -------------------------------------
             marshal    541.293k (±14.8%) i/s -      2.671M in   5.029854s
             msgpack    326.023k (±13.2%) i/s -      1.617M in   5.042620s
              bypass    664.092k (±20.6%) i/s -      3.252M in   5.068323s

Comparison:
             marshal:   541293.1 i/s
              bypass:   664092.5 i/s - same-ish: difference falls within error
             msgpack:   326023.2 i/s - 1.66x  (± 0.00) slower


 === Read 1000000B ===
Warming up --------------------------------------
             marshal   994.000  i/100ms
             msgpack   222.471k i/100ms
              bypass   525.644k i/100ms
Calculating -------------------------------------
             marshal     10.271k (± 4.9%) i/s -     51.688k in   5.044480s
             msgpack      2.218M (± 0.9%) i/s -     11.124M in   5.015168s
              bypass      5.226M (± 0.9%) i/s -     26.282M in   5.029194s

Comparison:
             marshal:    10271.0 i/s
              bypass:  5226377.5 i/s - 508.85x  (± 0.00) faster
             msgpack:  2218154.5 i/s - 215.96x  (± 0.00) faster

 === Write 1000000B ===
Warming up --------------------------------------
             marshal   887.000  i/100ms
             msgpack   470.000  i/100ms
              bypass   900.000  i/100ms
Calculating -------------------------------------
             marshal      9.442k (± 5.4%) i/s -     47.898k in   5.088189s
             msgpack      4.921k (± 3.8%) i/s -     24.910k in   5.068934s
              bypass      8.672k (± 6.2%) i/s -     43.200k in   5.002015s

Comparison:
             marshal:     9441.5 i/s
              bypass:     8671.8 i/s - same-ish: difference falls within error
             msgpack:     4921.4 i/s - 1.92x  (± 0.00) slower
```