@casperisfine 2022-11-28
Hum, I just realized that unfortunately this isn't really usable with `ActiveSupport::Cache`, because you'd usually serialize a `AS::Cache::Entry` which `Paquito::EntryCoder` turns into an `Array`.

That's really unfortunate. I'll see if I can figure something out.