@etiennebarrie 2022-11-28
That may not be desirable or doable, but I've noticed that discourse uses ActiveSupport::Cache without cache entries altogether. It also doesn't support versioning, and the expiry is done in Redis directly. But for a String, you may be able to skip the Array altogether in EntryCoder?