@casperisfine 2022-11-28
Unfortunately storing the TTL is necessary for `race_condition_ttl` which IMO is an important feature.

But I think we can handle both by moving the versioning lower.